+++ modules/angular2-express-engine/angular2-express-engine.d.ts
@@ -1,0 +1,7 @@
+export function ng2engineWithPreboot(): any;
+export function ng2engine(): any;
+
+export default {
+  ng2engineWithPreboot,
+  ng2engine
+}
