+++ modules/universal/server/server.ts
@@ -3,8 +3,6 @@ import './src/server_patch';
 export {provide} from 'angular2/core';
 export * from './src/directives/server_form';
 
-export * from './src/express/engine';
-
 export * from './src/http/server_http';
 
 export * from './src/platform/node';
@@ -15,5 +13,5 @@ export * from './src/router/server_router';
 
 export * from './src/helper';
 export * from './src/ng_preboot';
-export * from './src/ng_scripts';
+export * from './src/render';
 export * from './src/stringifyElement';
