+++ tsconfig.json
@@ -55,11 +55,9 @@
     "modules/universal/server/src/platform/dom/server_dom_renderer.ts",
     "modules/universal/server/src/platform/node.ts",
     "modules/universal/server/src/platform/node_xhr_impl.ts",
-    "modules/universal/server/src/express/engine.ts",
     "modules/universal/server/src/directives/server_form.ts",
     "modules/universal/server/src/http/server_http.ts",
     "modules/universal/server/src/router/server_router.ts",
-    "modules/universal/server/src/ng_scripts.ts",
     "modules/universal/server/src/helper.ts",
     "modules/universal/server/src/render.ts",
     "modules/universal/server/src/server_patch.ts",
@@ -68,6 +66,8 @@
     "modules/universal/client/index.ts",
     "modules/universal/client/client.ts",
     "modules/universal/client/src/ng_preload_cache.ts",
+    "modules/angular2-express-engine/index.ts",
+    "modules/angular2-express-engine/engine.ts",
     "examples/app/server/api.ts",
     "examples/app/server/routes.ts",
     "examples/app/server/server.ts",
@@ -80,7 +80,8 @@
     "examples/app/universal/test_router/client.ts"
   ],
   "exclude": [
-    "modules/universal/client/test",
+    "modules/angular2-express-engine/test",
+    "modules/angular2-express-engine/node_modules",
     "modules/universal/server/test",
     "modules/universal/node_modules",
     "preboot/test",
